refactor: Move `JsonContainer` conversion to dedicated module and traits.
